requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.11

# Caching
django-redis==5.4.0

# Background tasks
celery==5.3.4


# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# API Documentation
drf-spectacular==0.26.5

# File handling
Pillow==12.0.0

# Utilities
python-slugify==8.0.1
django-filter==23.3

# Development
django-debug-toolbar==4.2.0